feat(radio): integrate audio handler
